requirements.txt
Django==4.2
djangorestframework==3.15.0
sentence-transformers==2.2.2
chromadb==0.4.8
transformers==4.38.0
torch==2.2.0
pypdf==3.11.0
python-docx==0.8.11
gunicorn==21.2.0
uvicorn==0.21.1
python-dotenv==1.0.0
